Rename IpfsCatFunction.
